Move helpers and constants out to their own file
Makes the dependency graph a lot easier to tree-shake.
