+++ src/conversions/fromLMS.js
@@ -1,11 +1,11 @@
-const helpers = require('../helpers.js')
+const getTransform = require('../helpers/get-transform.js')
 
 module.exports = {
   XYZ: value => {
     let valueArray = [ value.rho, value.gamma, value.beta ]
 
       // Inverse Bradford Transformation
-    let Mbi = helpers.getTransform('INVERSE_BRADFORD')
+    let Mbi = getTransform('INVERSE_BRADFORD')
 
     let resultArray = Mbi.map((m) => {
       return valueArray.reduce((acc, v, key) => {
