+++ src/conversions/fromYiq.js
@@ -1,11 +1,11 @@
-const helpers = require('../helpers.js')
+const bounded = require('../helpers/bounded.js')
 
-const bound = v => helpers.bounded(v, [ 0, 255 ])
+const bound = v => bounded(v, [ 0, 255 ])
 
 module.exports = {
   rgb: value => {
-    const i = helpers.bounded(value.i, [ -0.5957, 0.5957 ])
-    const q = helpers.bounded(value.q, [ -0.5226, 0.5226 ])
+    const i = bounded(value.i, [ -0.5957, 0.5957 ])
+    const q = bounded(value.q, [ -0.5226, 0.5226 ])
 
     const r = 255 * bound(value.y + (0.956 * i) + (0.621 * q))
     const g = 255 * bound(value.y + (-0.272 * i) + (-0.647 * q))
