+++ src/helpers/get-illuminant.js
@@ -1,0 +1,5 @@
+const ILLUMINANTS = require('../constants/illuminants.js')
+
+module.exports = function getIlluminant (ref) {
+  return ILLUMINANTS[ref]
+}
