+++ src/helpers/get-transform.js
@@ -1,0 +1,5 @@
+const TRANSFORMS = require('../constants/transforms.js')
+
+module.exports = function getTransform (ref) {
+  return TRANSFORMS[ref]
+}
