+++ src/helpers/matrix-multiply.js
@@ -1,0 +1,14 @@
+module.exports = function matrixMultiply (a, b) {
+  var result = []
+  for (let i = 0; i < a.length; i++) {
+    result[i] = []
+    for (let j = 0; j < b[0].length; j++) {
+      var sum = 0
+      for (let k = 0; k < a[0].length; k++) {
+        sum += a[i][k] * b[k][j]
+      }
+      result[i][j] = sum
+    }
+  }
+  return result
+}
