+++ src/helpers/negative-modulo.js
@@ -1,0 +1,3 @@
+module.exports = function negativeModulo (n, m) {
+  return ((n % m) + m) % m
+}
