+++ src/helpers/to-degree.js
@@ -1,0 +1,3 @@
+module.exports = function toDegree (angle) {
+  return angle * (180 / Math.PI)
+}
