+++ src/helpers/to-radian.js
@@ -1,0 +1,3 @@
+module.exports = function toRadian (angle) {
+  return angle * (Math.PI / 180)
+}
