+++ src/operations/adapt.js
@@ -1,4 +1,6 @@
-const { getIlluminant, matrixMultiply, getTransform } = require('../helpers')
+const getIlluminant = require('../helpers/get-illuminant.js')
+const matrixMultiply = require('../helpers/matrix-multiply.js')
+const getTransform = require('../helpers/get-transform.js')
 const convert = require('../helpers/convert-to-type.js')
 const makeColourObject = require('./convert.js')
