+++ src/operations/adjacent.js
@@ -1,4 +1,4 @@
-const { negMod } = require('../helpers')
+const negativeModulo = require('../helpers/negative-modulo.js')
 const convert = require('../helpers/convert-to-type.js')
 const makeColourObject = require('./convert.js')
 
@@ -7,7 +7,7 @@ function adjacent (deg, amount, colourRef) {
   const colours = [{ h: colour.h, s: colour.s, l: colour.l }]
 
   for (let i = 0; i < (amount - 1); i++) {
-    colour.h = negMod((colour.h + deg), 360)
+    colour.h = negativeModulo((colour.h + deg), 360)
     colours.push({ h: colour.h, s: colour.s, l: colour.l })
   }
